Added ability to fetch from a specified set of farend files rather than just one file
The list_of_wav_filenames.json specifies a list of wav files that are available for download to the local device. This json file is currently hand edited and is NOT (currently) automatically generated.
This list of filenames is fetched to the local device and used to populate the dropdown box of wav files.
The specified wav file must be loaded (by pressing Load button) before it can be played over the local loudspeakers
Also added a Record + Play button so that the local device starts recording AND THEN the downloaded farend file starts playback.